Tweaks
Made the trigger guard slightly thicker and should actually be able to be put in place now. One end requires a nut and a M3 screw.

Added a medium muzzle for those who want short but don't want the tight inner diameter of the 19mmOD Muzzle.

Changed the screw hole for the split upper rear to use the same method as the other screw holes.